requirements.txt
annotated-types==0.6.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.1.1
cycler==0.12.1
Cython==0.29.14
DateTime==5.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==4.2.9
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.8.0
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.2.0
gensim==3.8.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.0
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lesscpy==0.15.1
MarkupSafe==2.1.3
matplotlib==3.7.5
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.24.4
openai==1.12.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.1.0
plotly==5.19.0
ply==3.11
praw==7.7.1
prawcore==2.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
sympy==1.12
tenacity==8.2.3
tensorly==0.8.1
terminado==0.18.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.1+cpu
torchaudio==2.2.1+cpu
torchvision==0.17.1+cpu
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
unicode==2.9
Unidecode==1.3.8
update-checker==0.18.0
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
zipp==3.17.0
zope.interface==6.2
